requirements.txt
absl-py==2.1.0
accelerate==1.5.2
addict==2.4.0
affine==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
airsim==1.8.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.2.0
babel==2.17.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.3
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
cityscapesscripts==2.2.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.2
datasets==3.4.1
deepspeed==0.16.2
decorator==5.1.1
defusedxml==0.7.1
diffdist==0.1
dill==0.3.8
distro==1.9.0
django==5.1.7
docker-pycreds==0.4.0
docstring-parser==0.16
einops==0.8.1
einops-exts==0.0.4
exceptiongroup==1.2.2
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
flash-attn==2.7.4.post1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.21.0
gradio-client==1.7.2
groovy==0.1.2
grpcio==1.68.1
gymnasium==1.0.0
h11==0.14.0
hf-transfer==0.1.9
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
image==1.5.33
imageio==2.36.1
importlib-resources==6.4.5
infinibatch==0.1.1
iniconfig==2.1.0
inquirerpy==0.3.4
iopath==0.1.9
ipykernel
ipython
isort==6.0.1
jedi
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json-tricks==3.17.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client
jupyter-core
kiwisolver==1.4.8
kornia==0.6.8
latex2mathml==3.77.0
latex2sympy2-extended==1.10.1
Levenshtein==0.27.1
librosa==0.11.0
llvmlite==0.43.0
markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline
mccabe==0.7.0
mdurl==0.1.2
modelscope==1.25.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
multidict==6.2.0
multiprocess==0.70.16
MultiScaleDeformableAttention==1.0
mup==1.0.0
mypy-extensions==1.0.0
narwhals==1.20.1
networkx==3.4.2
ninja==1.11.1.4
nltk==3.8.1
numba==0.60.0
numpy==2.2.4
nvidia-ml-py==12.560.30
omegaconf==2.3.0
openai==1.68.2
openai-whisper==20240930
opencv-contrib-python==4.10.0.84
opencv-python==4.8.1.78
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso
pathspec==0.12.1
pathtools==0.1.2
peft==0.15.0
pexpect
pickleshare
Pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
portalocker==3.1.1
prompt-toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
ptyprocess
pure-eval
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycocotools==2.0.8
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydot==3.0.3
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq
RapidFuzz==3.12.2
rasterio==1.4.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge-chinese==1.0.3
rpds-py==0.22.3
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.24.0
setproctitle==1.3.5
shapely==1.8.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sqlparse==0.5.3
sse-starlette==2.2.1
stack-data
starlette==0.41.3
supervision==0.25.1
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.8.0
timm==0.6.13
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.6.0
torchvision==0.21.0
tornado==4.5.3
tqdm==4.67.1
traitlets
transformers==4.50.0
triton==3.2.0
trl==0.9.6
typer==0.15.1
typing==3.7.4.3
typing-extensions==4.12.2
tyro==0.8.14
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
vision-datasets==0.2.2
wandb==0.18.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zipp

# Git & Editable installs
-e git+https://github.com/om-ai-lab/VLM-R1.git@261f30a737f55cc472666d88a5a2288f4bf9516c#egg=open_r1&subdirectory=src/open-r1-multimodal
-e git+https://github.com/MaureenZOU/detectron2-xyz@42121d75e10d9f858f3a91b6a39f5722c02868f0
-e git+https://github.com/IDEA-Research/GroundingDINO.git@856dde20aee659246248e20734ef9ba5214f5e44
-e git+https://github.com/hiyouga/LLaMA-Factory.git@01166841cfe118b8792161a43af4db416835073e
-e git+https://github.com/water-cookie/LLaVA.git@c17aca33b353c3b15ea463c08b1f19a94d0f54c6
-e git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
-e git+https://github.com/water-cookie/Semantic-SAM.git@88d8cd66f7061b6bb72f07edce9a33e24b78e028
